refactor: split runs.rs god-file into modules
